refactor(F4): extract code action builders to features/code_actions.rs
Move is_non_call_keyword, build_introduce_variable, build_import_alias_action,
build_rename_placeholder_action, expand_call_expr, derive_var_name from
backend/actions.rs → features/code_actions.rs.

Move whole_word_replace_file from backend/rename.rs → features/text_utils.rs
(shared with features/code_actions.rs).

Add compute_code_actions() entry-point: pure fn taking text/uri/range, no
indexer access.  backend/actions.rs now a thin delegate (~50 LoC).
backend/rename.rs imports is_non_call_keyword from features::code_actions.
